Add simple feature classifier + bug fixes.
Added a classifier that matches the replay datta to the closest matching
player profile, the features are scaled between 0 and 1 and the L2
distance is used. Significant bug forced n_gram classifier to not
classify an account as itself, which makes sense for barcodes but not when testing the accuracy of it with a known player. This if fixed now.