Add new feature files to database
Rather than just saving the feature mean for each player we now also save the standard deviation as well as the average mean and standard deviation over all players (sum over all players of their mean and standard deviations and then divide by the number of players). This is aimed to help a future Bayesian-like classifier.